scheduler/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
apscheduler==3.9.1
    # via -r requirements.in
dependency-injector==4.40.0
    # via -r requirements.in
faker==15.0.0
    # via -r requirements.in
greenlet==1.1.3
    # via sqlalchemy
grpc-interceptor==0.15.0
    # via -r requirements.in
grpcio==1.49.1
    # via
    #   -r requirements.in
    #   grpc-interceptor
    #   grpcio-tools
grpcio-tools==1.49.1
    # via -r requirements.in
pika==1.3.0
    # via -r requirements.in
protobuf==4.21.6
    # via grpcio-tools
psycopg2-binary==2.9.3
    # via -r requirements.in
pydantic==1.10.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
pytz==2022.2.1
    # via apscheduler
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
six==1.16.0
    # via
    #   apscheduler
    #   dependency-injector
    #   grpcio
    #   python-dateutil
sqlalchemy==1.4.41
    # via -r requirements.in
typing-extensions==4.3.0
    # via pydantic
tzdata==2022.4
    # via pytz-deprecation-shim
tzlocal==4.2
    # via apscheduler

# The following packages are considered to be unsafe in a requirements file:
# setuptools
